+++ lib/td_dd_web/router.ex
@@ -59,6 +59,7 @@ defmodule TdDdWeb.Router do
     end
 
     get("/data_structures/search/reindex_all", SearchController, :reindex_all)
+    get("/data_structures/search/source_alias", SearchController, :get_source_aliases)
 
     get("/data_structure_filters", DataStructureFilterController, :index)
     post("/data_structure_filters/search", DataStructureFilterController, :search)