+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- [TD-2633] Allow logic deletion of rule implementations with results
+
 ## [3.22.0] 2020-05-18
 
 ### Added