+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Added
+
+- [TD-2562] Endpoint `GET /api/data_structures/search/source_alias` to return the list of 
+distinct structures metadata aliases
+
 ### Changed
 
 - [TD-2643] Show metadata on structure relations