+++ CHANGELOG.md
@@ -12,6 +12,7 @@ distinct structures metadata aliases
 - [TD-2643] Show metadata on structure relations
 - [TD-2487] Exclude deleted structures from csv download
 - [TD-2629] Update td-df-lib to omit template fields of type `image` on indexing
+- [TD-2492] Update td-df-lib to include new numeric template types
 
 ## [3.22.0] 2020-05-18