+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - [TD-2629] Update td-df-lib to omit template fields of type `image` on indexing
+- [TD-2492] Update td-df-lib to include new numeric template types
 
 ### Fixed